reorg structs and func along for each stat into one file
