+++ collector/pod_kubelet_ack_to_container_start_time_test.go
@@ -10,54 +10,6 @@ import (
 	"time"
 )
 
-func TestCreateKubeletLatencyFilter_Update(t *testing.T) {
-	filter := &createKubeletLatencyFilter{
-		metric: NewPodCreateToKubeletDurationMetric(),
-	}
-	for _, tc := range []struct {
-		name           string
-		oldPod         *corev1.Pod
-		newPod         *corev1.Pod
-		expectedRC     bool
-		expectedMetric bool
-	}{
-		{
-			name:   "both start times nil",
-			oldPod: &corev1.Pod{},
-			newPod: &corev1.Pod{},
-		},
-		{
-			name:           "only new start time set",
-			expectedMetric: true,
-			oldPod:         &corev1.Pod{},
-			newPod: &corev1.Pod{
-				Status: corev1.PodStatus{StartTime: &metav1.Time{Time: time.Now()}},
-			},
-		},
-		{
-			name: "both start times set",
-			oldPod: &corev1.Pod{
-				Status: corev1.PodStatus{StartTime: &metav1.Time{Time: time.Now()}},
-			},
-			newPod: &corev1.Pod{
-				Status: corev1.PodStatus{StartTime: &metav1.Time{Time: time.Now()}},
-			},
-		},
-	} {
-		ev := event.UpdateEvent{
-			ObjectOld: tc.oldPod,
-			ObjectNew: tc.newPod,
-		}
-		rc := filter.Update(ev)
-		if rc != tc.expectedRC {
-			t.Errorf(fmt.Sprintf("tc %s expected %v but got %v", tc.name, tc.expectedRC, rc))
-		}
-		if tc.expectedMetric {
-			validateHistogramVec(t, filter.metric, prometheus.Labels{NS_LABEL: tc.newPod.Namespace, TASK_NAME_LABEL: taskRef(tc.newPod.Labels)}, false)
-		}
-	}
-}
-
 func TestKubeletContainerLatencyFilter_Update(t *testing.T) {
 	filter := &kubeletContainerLatencyFilter{
 		metric: NewPodKubeletToContainerStartDurationMetric(),
@@ -192,6 +144,142 @@ func TestKubeletContainerLatencyFilter_Update(t *testing.T) {
 		}
 		if tc.expectedMetric {
 			validateHistogramVec(t, filter.metric, prometheus.Labels{NS_LABEL: tc.newPod.Namespace, TASK_NAME_LABEL: taskRef(tc.newPod.Labels)}, false)
+		}
+	}
+}
+
+func TestCalculateTaskRunPodStartedToFirstContainerStartedDuration(t *testing.T) {
+	now := time.Now()
+	for _, tc := range []struct {
+		expectedAmt float64
+		pod         *corev1.Pod
+	}{
+		{
+			expectedAmt: 0,
+			pod: &corev1.Pod{
+				ObjectMeta: metav1.ObjectMeta{
+					Name:              "test-taskrun-1",
+					Namespace:         "test-namespace",
+					CreationTimestamp: metav1.NewTime(now),
+				},
+			},
+		},
+		{
+			expectedAmt: 0,
+			pod: &corev1.Pod{
+				ObjectMeta: metav1.ObjectMeta{
+					Name:              "test-taskrun-1",
+					Namespace:         "test-namespace",
+					CreationTimestamp: metav1.NewTime(now),
+				},
+				Status: corev1.PodStatus{StartTime: &metav1.Time{Time: now.Add(3 * time.Second)}},
+			},
+		},
+		{
+			expectedAmt: 0,
+			pod: &corev1.Pod{
+				ObjectMeta: metav1.ObjectMeta{
+					Name:              "test-taskrun-1",
+					Namespace:         "test-namespace",
+					CreationTimestamp: metav1.NewTime(now),
+				},
+				Status: corev1.PodStatus{StartTime: &metav1.Time{Time: now.Add(3 * time.Second)}},
+			},
+		},
+		{
+			expectedAmt: 2000,
+			pod: &corev1.Pod{
+				ObjectMeta: metav1.ObjectMeta{
+					Name:              "test-taskrun-1",
+					Namespace:         "test-namespace",
+					CreationTimestamp: metav1.NewTime(now),
+				},
+				Status: corev1.PodStatus{
+					StartTime: &metav1.Time{Time: now.Add(3 * time.Second)},
+					ContainerStatuses: []corev1.ContainerStatus{
+						{
+							State: corev1.ContainerState{
+								Running: &corev1.ContainerStateRunning{StartedAt: metav1.Time{Time: now.Add(5 * time.Second)}},
+							},
+						},
+					},
+				},
+			},
+		},
+		{
+			expectedAmt: 2000,
+			pod: &corev1.Pod{
+				ObjectMeta: metav1.ObjectMeta{
+					Name:              "test-taskrun-1",
+					Namespace:         "test-namespace",
+					CreationTimestamp: metav1.NewTime(now),
+				},
+				Status: corev1.PodStatus{
+					StartTime: &metav1.Time{Time: now.Add(3 * time.Second)},
+					ContainerStatuses: []corev1.ContainerStatus{
+						{
+							State: corev1.ContainerState{
+								Terminated: &corev1.ContainerStateTerminated{StartedAt: metav1.Time{Time: now.Add(5 * time.Second)}},
+							},
+						},
+					},
+				},
+			},
+		},
+		{
+			expectedAmt: 2000,
+			pod: &corev1.Pod{
+				ObjectMeta: metav1.ObjectMeta{
+					Name:              "test-taskrun-1",
+					Namespace:         "test-namespace",
+					CreationTimestamp: metav1.NewTime(now),
+				},
+				Status: corev1.PodStatus{
+					StartTime: &metav1.Time{Time: now.Add(3 * time.Second)},
+					ContainerStatuses: []corev1.ContainerStatus{
+						{
+							State: corev1.ContainerState{
+								Running: &corev1.ContainerStateRunning{StartedAt: metav1.Time{Time: now.Add(6 * time.Second)}},
+							},
+						},
+						{
+							State: corev1.ContainerState{
+								Terminated: &corev1.ContainerStateTerminated{StartedAt: metav1.Time{Time: now.Add(5 * time.Second)}},
+							},
+						},
+					},
+				},
+			},
+		},
+		{
+			expectedAmt: 2000,
+			pod: &corev1.Pod{
+				ObjectMeta: metav1.ObjectMeta{
+					Name:              "test-taskrun-1",
+					Namespace:         "test-namespace",
+					CreationTimestamp: metav1.NewTime(now),
+				},
+				Status: corev1.PodStatus{
+					StartTime: &metav1.Time{Time: now.Add(3 * time.Second)},
+					ContainerStatuses: []corev1.ContainerStatus{
+						{
+							State: corev1.ContainerState{
+								Running: &corev1.ContainerStateRunning{StartedAt: metav1.Time{Time: now.Add(5 * time.Second)}},
+							},
+						},
+						{
+							State: corev1.ContainerState{
+								Terminated: &corev1.ContainerStateTerminated{StartedAt: metav1.Time{Time: now.Add(6 * time.Second)}},
+							},
+						},
+					},
+				},
+			},
+		},
+	} {
+		got := calculateTaskRunPodStartedToFirstContainerStartedDuration(tc.pod)
+		if got != tc.expectedAmt {
+			t.Errorf("expected %v but got %v", tc.expectedAmt, got)
 		}
 	}
 }
